+++ src/main/java/com/zufar/icedlatte/review/dto/ProductRatingCount.java
@@ -1,3 +1,3 @@
 package com.zufar.icedlatte.review.dto;
 
-public record ProductRatingCount(int productRating, int count) {}
+public record ProductRatingCount(int productRating, long count) {}
